스크립트 정리 및 적 버그 수정
+ 적이 게임 밖에 벗어나면 점수를 얻는 버그 수정

+++ c/Games/ConsoleGame1/ConsoleGame1/DataTypes.cpp
@@ -27,11 +27,11 @@ int GetScreenPositionByRatio(int type, float value)
 {
 	if (type == 0) // 가로
 	{
-		return MAXWIDTH * value;
+		return (int)(MAXWIDTH * value);
 	}
 	else if (type == 1) // 세로
 	{
-		return MAXHEIGHT * value;
+		return (int)(MAXHEIGHT * value);
 	}
 
 	return 0; // 잘못된 파라미터 값
